requirements.txt
anyio==4.4.0
APScheduler==3.6.3
asgiref==3.8.1
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-dotenv==1.0.1
python-telegram-bot==21.2
sniffio==1.3.1
cachetools==4.2.2
certifi==2024.2.2
Django==5.0.6
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
psycopg2==2.9.9
python-dotenv==1.0.1
python-telegram-bot==21.2
pytz==2024.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
uvicorn==0.23.2
tornado==6.4
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2